feat: split notification editor to multiple files
